# smv2/openapi.json
{
  "components": {
    "schemas": {
...
      "AdaptiveStudyQueueOut": {
        "properties": {
          "activities": {
            "items": {
              "$ref": "#/components/schemas/AdaptiveStudyActivityOut"
            },
            "title": "Activities",
            "type": "array"
          }
        },
        "required": [
          "activities"
        ],
        "title": "AdaptiveStudyQueueOut",
        "type": "object"
      },
...
      "CourseOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failed_asset_count": {
            "default": 0,
            "title": "Failed Asset Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProgressSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "section_count": {
            "default": 0,
            "title": "Section Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "CourseOut",
        "type": "object"
      },
...
      "GradeCardIn": {
        "properties": {
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "grade": {
            "title": "Grade",
            "type": "integer"
          }
        },
        "required": [
          "grade"
        ],
        "title": "GradeCardIn",
        "type": "object"
      },
      "GradeCardOut": {
        "properties": {
          "next_due_at": {
            "format": "date-time",
            "title": "Next Due At",
            "type": "string"
          },
          "remaining_due": {
            "title": "Remaining Due",
            "type": "integer"
          }
        },
        "required": [
          "next_due_at",
          "remaining_due"
        ],
        "title": "GradeCardOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RetentionAssignmentIn": {
        "properties": {
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "workload_target": {
            "default": 12,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Workload Target",
            "type": "integer"
          }
        },
        "required": [
          "concept_id"
        ],
        "title": "RetentionAssignmentIn",
        "type": "object"
      },
      "RetentionAssignmentOut": {
        "properties": {
          "assigned_at": {
            "format": "date-time",
            "title": "Assigned At",
            "type": "string"
          },
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "study_group": {
            "title": "Study Group",
            "type": "string"
          },
          "study_id": {
            "title": "Study Id",
            "type": "string"
          },
          "workload_target": {
            "title": "Workload Target",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "study_id",
          "concept_id",
          "study_group",
          "workload_target",
          "assigned_at"
        ],
        "title": "RetentionAssignmentOut",
        "type": "object"
      },
      "RetentionProbeIn": {
        "properties": {
          "learning_claim_id": {
            "title": "Learning Claim Id",
            "type": "string"
          }
        },
        "required": [
          "learning_claim_id"
        ],
        "title": "RetentionProbeIn",
        "type": "object"
      },
      "RetentionProbeOut": {
        "properties": {
          "assignment_id": {
            "title": "Assignment Id",
            "type": "string"
          },
          "evidence_item_id": {
            "title": "Evidence Item Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "learning_claim_id": {
            "title": "Learning Claim Id",
            "type": "string"
          },
          "scheduled_for": {
            "format": "date-time",
            "title": "Scheduled For",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "assignment_id",
          "evidence_item_id",
          "learning_claim_id",
          "scheduled_for",
          "status"
        ],
        "title": "RetentionProbeOut",
        "type": "object"
      },
      "RetentionStudyIn": {
        "properties": {
          "assignment_seed": {
            "title": "Assignment Seed",
            "type": "string"
          },
          "delay_end_days": {
            "default": 14,
            "minimum": 1.0,
            "title": "Delay End Days",
            "type": "integer"
          },
          "delay_start_days": {
            "default": 7,
            "minimum": 1.0,
            "title": "Delay Start Days",
            "type": "integer"
          },
          "minimum_per_group": {
            "default": 20,
            "minimum": 1.0,
            "title": "Minimum Per Group",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "assignment_seed"
        ],
        "title": "RetentionStudyIn",
        "type": "object"
      },
      "RetentionStudyOut": {
        "properties": {
          "assignment_seed": {
            "title": "Assignment Seed",
            "type": "string"
          },
          "course_id": {
            "title": "Course Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delay_end_days": {
            "title": "Delay End Days",
            "type": "integer"
          },
          "delay_start_days": {
            "title": "Delay Start Days",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "minimum_per_group": {
            "title": "Minimum Per Group",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "protocol_version": {
            "title": "Protocol Version",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "course_id",
          "name",
          "status",
          "assignment_seed",
          "protocol_version",
          "delay_start_days",
          "delay_end_days",
          "minimum_per_group",
          "created_at"
        ],
        "title": "RetentionStudyOut",
        "type": "object"
      },
...
      "ReviewQueueOut": {
        "properties": {
          "cards": {
            "items": {
              "$ref": "#/components/schemas/ReviewQueueCardOut"
            },
            "title": "Cards",
            "type": "array"
          },
          "due": {
            "title": "Due",
            "type": "integer"
          },
          "new": {
            "title": "New",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "cards",
          "due",
          "new",
          "total"
        ],
        "title": "ReviewQueueOut",
        "type": "object"
      },
...
      "SkillDetailOut": {
        "properties": {
          "cards_count": {
            "title": "Cards Count",
            "type": "integer"
          },
          "missed_questions": {
            "items": {
              "$ref": "#/components/schemas/SkillMissedQuestionOut"
            },
            "title": "Missed Questions",
            "type": "array"
          },
          "node": {
            "$ref": "#/components/schemas/SkillNodeOut"
          },
          "quiz_correct": {
            "title": "Quiz Correct",
            "type": "integer"
          },
          "quiz_wrong": {
            "title": "Quiz Wrong",
            "type": "integer"
          },
          "taught_in": {
            "items": {
              "$ref": "#/components/schemas/SkillTaughtInOut"
            },
            "title": "Taught In",
            "type": "array"
          }
        },
        "required": [
          "node",
          "taught_in",
          "missed_questions",
          "cards_count",
          "quiz_correct",
          "quiz_wrong"
        ],
        "title": "SkillDetailOut",
        "type": "object"
      },
      "SkillEdgeOut": {
        "properties": {
          "from_id": {
            "title": "From Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "to_id": {
            "title": "To Id",
            "type": "string"
          }
        },
        "required": [
          "from_id",
          "to_id",
          "kind"
        ],
        "title": "SkillEdgeOut",
        "type": "object"
      },
...
      "SkillGraphImportOut": {
        "properties": {
          "concept_count": {
            "title": "Concept Count",
            "type": "integer"
          },
          "edge_count": {
            "title": "Edge Count",
            "type": "integer"
          },
          "link_count": {
            "title": "Link Count",
            "type": "integer"
          }
        },
        "required": [
          "concept_count",
          "edge_count",
          "link_count"
        ],
        "title": "SkillGraphImportOut",
        "type": "object"
      },
      "SkillGraphIn": {
        "properties": {
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/SkillGraphConceptIn"
            },
            "maxItems": 500,
            "title": "Concepts",
            "type": "array"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/SkillGraphEdgeIn"
            },
            "maxItems": 2000,
            "title": "Edges",
            "type": "array"
          }
        },
        "required": [
          "concepts"
        ],
        "title": "SkillGraphIn",
        "type": "object"
      },
...
      "SkillMapOut": {
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/SkillEdgeOut"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/SkillNodeOut"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "required": [
          "nodes",
          "edges"
        ],
        "title": "SkillMapOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SourceMind v2 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cards/{card_id}/grade": {
      "post": {
        "operationId": "grade_card",
        "parameters": [
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "schema": {
              "title": "Card Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeCardIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeCardOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grade Card",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/courses/{course_id}": {
      "delete": {
        "operationId": "delete_course",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Course",
        "tags": [
          "courses"
        ]
      },
      "get": {
        "operationId": "get_course",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Course",
        "tags": [
          "courses"
        ]
      }
    },
...
    "/api/courses/{course_id}/retention-studies": {
      "post": {
        "operationId": "create_retention_study",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionStudyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionStudyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Retention Study",
        "tags": [
          "retention-studies"
        ]
      }
    },
    "/api/courses/{course_id}/retention-studies/{study_id}/assignments": {
      "post": {
        "operationId": "create_retention_assignment",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionAssignmentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionAssignmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Retention Assignment",
        "tags": [
          "retention-studies"
        ]
      }
    },
    "/api/courses/{course_id}/retention-studies/{study_id}/assignments/{assignment_id}/probes": {
      "post": {
        "operationId": "schedule_retention_probe",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionProbeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionProbeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schedule Retention Probe",
        "tags": [
          "retention-studies"
        ]
      }
    },
    "/api/courses/{course_id}/review/queue": {
      "get": {
        "operationId": "review_queue",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewQueueOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Queue",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/courses/{course_id}/skills": {
      "get": {
        "operationId": "get_skill_map",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillMapOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Skill Map",
        "tags": [
          "skills"
        ]
      }
    },
    "/api/courses/{course_id}/skills/graph": {
      "put": {
        "operationId": "import_skill_graph",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillGraphIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillGraphImportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Skill Graph",
        "tags": [
          "skills"
        ]
      }
    },
    "/api/courses/{course_id}/skills/{concept_id}": {
      "get": {
        "operationId": "get_skill_detail",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "concept_id",
            "required": true,
            "schema": {
              "title": "Concept Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Skill Detail",
        "tags": [
          "skills"
        ]
      }
    },
...
    "/api/courses/{course_id}/study/concepts/{concept_id}/replenish": {
      "post": {
        "operationId": "replenish_concept_practice",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "concept_id",
            "required": true,
            "schema": {
              "title": "Concept Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replenish Concept Practice",
        "tags": [
          "study"
        ]
      }
    },
    "/api/courses/{course_id}/study/queue": {
      "get": {
        "operationId": "adaptive_study_queue",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdaptiveStudyQueueOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adaptive Study Queue",
        "tags": [
          "study"
        ]
      }
    },
...
  }
}